Various improvements to the captive portal :  - translation now occurs within html templates, using the new function pf::web::i18n  - improved error handling when sending emails (email_activation.pm, sms_activation.pm)  - updated English localization file  - minor CSS/HTML fixes
Monotone-Parent: 6ae15136c013af0ab1a699b010d915f51dc39af6
Monotone-Revision: 26a5ad461dc0146ce42fe40b32296c7e66b17f97

Monotone-Author: flachapelle@inverse.ca
Monotone-Date: 2011-08-24T20:15:05
Monotone-Branch: org.packetfence.trunk

+++ pf/html/captive-portal/guest-management.cgi
@@ -61,7 +61,7 @@ foreach my $param($cgi->param()) {
 # is user already logged in?
 if (defined($session->param("login"))) {
 
-    if ($cgi->param("action") eq "logout") {
+    if (defined($cgi->param("action")) && $cgi->param("action") eq "logout") {
         $session->delete();
         pf::web::guest::generate_activation_login_page($cgi, $session, 0, "guest/mgmt_login.html");
         exit(0);
